refactor: LEMON - update classes, packages
