+++ piikii-application/src/main/kotlin/com/piikii/application/port/output/web/PlaceAutoCompleteClient.kt
@@ -2,6 +2,6 @@ package com.piikii.application.port.output.web
 
 import com.piikii.application.domain.place.OriginPlace
 
-interface UrlClient {
-    fun get(url: String): OriginPlace
+interface PlaceAutoCompleteClient {
+    fun getAutoCompletedPlace(url: String): OriginPlace?
 }
