+++ piikii-bootstrap/src/main/kotlin/com/piikii/bootstrap/PiikiiBootstrapApplication.kt
@@ -5,7 +5,7 @@ import org.springframework.boot.context.properties.ConfigurationPropertiesScan
 import org.springframework.boot.runApplication
 
 @SpringBootApplication(scanBasePackages = ["com.piikii"])
-@ConfigurationPropertiesScan("com.piikii.output.storage.ncp.config", "com.piikii.output.web.lemon.config")
+@ConfigurationPropertiesScan("com.piikii.output.storage.ncp.config")
 class PiikiiBootstrapApplication
 
 fun main(args: Array<String>) {
