+++ piikii-output-web/build.gradle.kts
@@ -1,17 +1,16 @@
 project(":piikii-output-web:avocado") {
     dependencies {
-        implementation("org.springframework:spring-web")
     }
 }
 
 project(":piikii-output-web:lemon") {
     dependencies {
-        implementation("org.springframework:spring-web")
     }
 }
 
 allprojects {
     dependencies {
+        implementation("org.springframework:spring-web")
         implementation(project(":piikii-application"))
     }
