+++ piikii-output-web/lemon/src/main/kotlin/com/piikii/output/web/lemon/adapter/LemonPlaceAutoCompleteClient.kt
@@ -3,21 +3,20 @@ package com.piikii.output.web.lemon.adapter
 import com.fasterxml.jackson.databind.ObjectMapper
 import com.fasterxml.jackson.module.kotlin.readValue
 import com.piikii.application.domain.place.OriginPlace
-import com.piikii.application.port.output.web.UrlClient
+import com.piikii.application.port.output.web.PlaceAutoCompleteClient
 import com.piikii.common.exception.ExceptionCode
 import com.piikii.common.exception.PiikiiException
-import com.piikii.output.web.lemon.UrlClientResponse
-import com.piikii.output.web.lemon.util.SaltAdditive
+import com.piikii.output.web.lemon.salt.LemonSaltAdditive
 import org.springframework.stereotype.Component
 import org.springframework.web.client.RestClient
 
 @Component
-class LemonUrlClient(
-    private val saltAdditive: SaltAdditive,
+class LemonPlaceAutoCompleteClient(
+    private val lemonSaltAdditive: LemonSaltAdditive,
     private val objectMapper: ObjectMapper,
-) : UrlClient {
-    override fun get(url: String): OriginPlace {
-        val saltedUrl = saltAdditive.execute(url)
+) : PlaceAutoCompleteClient {
+    override fun getAutoCompletedPlace(url: String): OriginPlace? {
+        val saltedUrl = lemonSaltAdditive.execute(url)
         val response = fetchResponse(saltedUrl)
         val placeResponse = parseResponse(response)
         return placeResponse.toOriginPlace(url)
@@ -31,7 +30,7 @@ class LemonUrlClient(
         )
     }
 
-    private fun parseResponse(response: String): UrlClientResponse {
+    private fun parseResponse(response: String): LemonPlaceInfoResponse {
         return objectMapper.readValue(response)
     }
 }
